feat(sglang): support GLM 5.2 MTP CUDA graphs
