+++ atom/plugin/sglang/attention_backend/glm52_dsa_backend.py
@@ -2,16 +2,119 @@
 
 from __future__ import annotations
 
+import copy
 import logging
-import os
 from types import SimpleNamespace
 
 import torch
 from sglang.srt.layers.attention.base_attn_backend import AttentionBackend
 
-from atom.plugin.sglang.glm52_mtp.common import GLM52_GRAPH_SEQ_LEN_CAPACITY
+from atom.plugin.sglang.glm52_dsa_bridge import GLM52_GRAPH_SEQ_LEN_CAPACITY
 
 logger = logging.getLogger("atom.plugin.sglang.attention_backend.glm52_dsa")
+
+
+def install_upstream_glm52_graph_metadata_adapter() -> None:
+    """Publish ATOM target-verify metadata from SGLang's active DSA backend."""
+    from sglang.srt.layers.attention.dsa_backend import DeepseekSparseAttnBackend
+
+    if getattr(
+        DeepseekSparseAttnBackend,
+        "_atom_glm52_graph_metadata_patched",
+        False,
+    ):
+        return
+
+    original_out_graph = DeepseekSparseAttnBackend.init_forward_metadata_out_graph
+
+    def init_forward_metadata_out_graph(
+        self,
+        forward_batch,
+        in_capture: bool = False,
+    ):
+        original_out_graph(self, forward_batch, in_capture=in_capture)
+
+        forward_mode = getattr(forward_batch, "forward_mode", None)
+        if not bool(getattr(forward_mode, "is_target_verify", lambda: False)()):
+            self.atom_glm52_graph_metadata = None
+            return
+
+        from atom.config import get_current_atom_config
+        from atom.plugin.sglang.glm52_dsa_bridge import (
+            build_mtp_verify_decode_metadata,
+        )
+        from atom.plugin.sglang.runtime.model_arch import is_glm52_dsa_config
+
+        atom_config = get_current_atom_config()
+        if not is_glm52_dsa_config(getattr(atom_config, "hf_config", None)):
+            return
+
+        positions = getattr(forward_batch, "positions", None)
+        if not torch.is_tensor(positions):
+            raise RuntimeError(
+                "GLM target-verify graph metadata requires capture positions"
+            )
+
+        metadata_batch = forward_batch
+        metadata_positions = positions
+        if in_capture:
+            draft_token_num = int(
+                getattr(getattr(forward_batch, "spec_info", None), "draft_token_num", 0)
+                or 0
+            )
+            if draft_token_num <= 0:
+                raise RuntimeError(
+                    "GLM target-verify graph capture requires draft_token_num"
+                )
+            prefix_len = GLM52_GRAPH_SEQ_LEN_CAPACITY - draft_token_num
+            metadata_positions = (
+                torch.arange(
+                    draft_token_num,
+                    dtype=positions.dtype,
+                    device=positions.device,
+                ).repeat(int(forward_batch.batch_size))
+                + prefix_len
+            )
+            metadata_batch = copy.copy(forward_batch)
+            metadata_batch.positions = metadata_positions
+
+        staged = build_mtp_verify_decode_metadata(
+            metadata_batch,
+            metadata_positions,
+            token_to_kv_pool=self.token_to_kv_pool,
+            req_to_token_pool=self.req_to_token_pool,
+            atom_config=atom_config,
+            use_positions=True,
+        )
+        batch_size = int(forward_batch.batch_size)
+        fixed_by_bs = getattr(
+            self,
+            "_atom_glm52_target_verify_graph_metadata",
+            None,
+        )
+        if fixed_by_bs is None:
+            fixed_by_bs = {}
+            self._atom_glm52_target_verify_graph_metadata = fixed_by_bs
+
+        if in_capture:
+            fixed = staged
+            fixed_by_bs[batch_size] = fixed
+        else:
+            fixed = fixed_by_bs.get(batch_size)
+            if fixed is None:
+                raise RuntimeError(
+                    "Missing GLM target-verify graph metadata for "
+                    f"batch size {batch_size}"
+                )
+            ATOMGLM52DSABackendForSgl._copy_metadata_in_place(staged, fixed)
+
+        self.atom_glm52_graph_metadata = fixed
+        forward_batch.atom_glm52_graph_metadata = fixed
+
+    DeepseekSparseAttnBackend.init_forward_metadata_out_graph = (
+        init_forward_metadata_out_graph
+    )
+    DeepseekSparseAttnBackend._atom_glm52_graph_metadata_patched = True
 
 
 class ATOMGLM52DSABackendForSgl(AttentionBackend):
@@ -70,6 +173,27 @@ class ATOMGLM52DSABackendForSgl(AttentionBackend):
             target_tensor = getattr(target, name, None)
             if not torch.is_tensor(source_tensor) or not torch.is_tensor(target_tensor):
                 continue
+            if name == "block_tables":
+                if source_tensor.dim() != 2 or target_tensor.dim() != 2:
+                    raise RuntimeError(
+                        "GLM target-verify block_tables must be two-dimensional: "
+                        f"runtime={tuple(source_tensor.shape)}, "
+                        f"capture={tuple(target_tensor.shape)}"
+                    )
+                if (
+                    source_tensor.shape[0] > target_tensor.shape[0]
+                    or source_tensor.shape[1] > target_tensor.shape[1]
+                ):
+                    raise RuntimeError(
+                        "GLM target-verify graph field block_tables exceeds "
+                        f"capture capacity: runtime={tuple(source_tensor.shape)}, "
+                        f"capture={tuple(target_tensor.shape)}"
+                    )
+                target_tensor.zero_()
+                target_tensor[: source_tensor.shape[0], : source_tensor.shape[1]].copy_(
+                    source_tensor
+                )
+                continue
             if source_tensor.numel() > target_tensor.numel():
                 raise RuntimeError(
                     f"GLM target-verify graph field {name} exceeds capture capacity: "
@@ -90,7 +214,7 @@ class ATOMGLM52DSABackendForSgl(AttentionBackend):
         positions,
     ):
         from atom.config import get_current_atom_config
-        from atom.plugin.sglang.glm52_mtp.target_verify import (
+        from atom.plugin.sglang.glm52_dsa_bridge import (
             build_mtp_verify_decode_metadata,
         )
 
@@ -100,12 +224,13 @@ class ATOMGLM52DSABackendForSgl(AttentionBackend):
             token_to_kv_pool=self.token_to_kv_pool,
             req_to_token_pool=self.req_to_token_pool,
             atom_config=get_current_atom_config(),
+            use_positions=True,
         )
 
     def _publish_target_verify_graph_metadata(self, forward_batch, metadata):
         self.atom_glm52_graph_metadata = metadata
         self.forward_metadata = forward_batch
-        setattr(forward_batch, "atom_glm52_graph_metadata", metadata)
+        forward_batch.atom_glm52_graph_metadata = metadata
         ATOMGLM52DSABackendForSgl._last_atom_glm52_graph_metadata = metadata
         return metadata
 
@@ -195,7 +320,9 @@ class ATOMGLM52DSABackendForSgl(AttentionBackend):
                 draft_token_num,
             ) + torch.arange(
                 draft_token_num, dtype=torch.int64, device=self.device
-            ).repeat(int(bs))
+            ).repeat(
+                int(bs)
+            )
             forward_batch = SimpleNamespace(
                 forward_mode=forward_mode,
                 actual_forward_mode=forward_mode,
@@ -215,9 +342,7 @@ class ATOMGLM52DSABackendForSgl(AttentionBackend):
                 forward_batch, positions
             )
             self._target_verify_graph_metadata[int(bs)] = metadata
-            return self._publish_target_verify_graph_metadata(
-                forward_batch, metadata
-            )
+            return self._publish_target_verify_graph_metadata(forward_batch, metadata)
         forward_batch = SimpleNamespace(
             forward_mode=forward_mode,
             actual_forward_mode=forward_mode,
@@ -275,9 +400,7 @@ class ATOMGLM52DSABackendForSgl(AttentionBackend):
             positions = graph_buffers.positions[:total_tokens]
             staged_batch = SimpleNamespace(
                 forward_mode=forward_mode,
-                actual_forward_mode=getattr(
-                    replay_batch, "forward_mode", forward_mode
-                ),
+                actual_forward_mode=getattr(replay_batch, "forward_mode", forward_mode),
                 batch_size=int(bs),
                 req_pool_indices=req_pool_indices,
                 seq_lens=seq_lens,
@@ -286,9 +409,7 @@ class ATOMGLM52DSABackendForSgl(AttentionBackend):
                 positions=positions,
                 spec_info=SimpleNamespace(draft_token_num=draft_token_num),
             )
-            staged = self._build_target_verify_graph_metadata(
-                staged_batch, positions
-            )
+            staged = self._build_target_verify_graph_metadata(staged_batch, positions)
             self._copy_metadata_in_place(staged, fixed)
             return self._publish_target_verify_graph_metadata(staged_batch, fixed)
         forward_batch = SimpleNamespace(
@@ -351,7 +472,6 @@ class ATOMGLM52DSABackendForSgl(AttentionBackend):
             max_bs=bs,
         )
         del max_num_tokens
-        return None
 
     def get_cuda_graph_seq_len_fill_value(self):
         return int(self._cuda_graph_seq_len_fill_value)
